% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{angleCalc}
\alias{angleCalc}
\title{Compute angles between reservoir and conceptor space}
\usage{
angleCalc(RS, CRS)
}
\arguments{
\item{RS}{reservoir states}

\item{CRS}{conceptor states}
}
\value{
conputed angles
}
\description{
Compute angles between reservoir and conceptor space
}
