% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bootdata}
\alias{bootdata}
\title{Generate bootstrapped data}
\usage{
bootdata(data, washoutL, trainL, blockL)
}
\arguments{
\item{data}{A T\code{x}d data set with columns as variables.}

\item{washoutL}{Number of time points used for reservoir washout.}

\item{trainL}{Number of time points used for conceptor training.}

\item{blockL}{Number of time points for each block in bootstrap procedure.}
}
\value{
Bootstrapped data for MBB procedure of the same dimension as \code{data}.
}
\description{
Creates bootstrapped data for MBB procedure of the same dimension as the original data. Randomly appends data sections of a specified block length after a washout and training period.
}
