% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ccp}
\alias{ccp}
\title{Change Point Detection with Conceptors}
\usage{
ccp(
  data,
  washL_plus_trainL = "",
  trainL = "",
  washoutL = "",
  tol = 0.04,
  nboots = 200
)
}
\arguments{
\item{data}{A T\code{x}d data set with variables as columns.}

\item{washL_plus_trainL}{Number of time points used for both reservoir washout and conceptor training.}

\item{trainL}{Number of time points used for conceptor training.}

\item{washoutL}{Number of time points used for reservoir washout.}

\item{tol}{Error tolerance for conceptor fit to the data.}

\item{nboots}{Number of bootstraps to estimate statistic null distribution.}
}
\value{
List of output:
\describe{
\item{\code{estimate}}{Estimated change point location.}
\item{\code{statistic}}{Statistic from method.}
\item{\code{MBBsig}}{Significance estimate from MBB.}
\item{\code{MBBnull}}{Simulated null distribution from MBB.}
\item{\code{statSeries}}{Time-ordered series of statistics.}
\item{\code{angles}}{Time-ordered cosine similarities between reservoir states and conceptor space.}
\item{\code{netParams}}{List of RNN parameters:}
\itemize{
\item{\code{rscale}} {Scale of reservoir matrix.}
\item{\code{iscale}} {Scale of input matrix.}
\item{\code{bscale}} {Scale of bias matrix.}
\item{\code{N}} {RNN reservoir size.}
\item{\code{aperture} }{Parameter for conceptor computation.}
\item{\code{washoutL}} {Number of time points used for reservoir washout.}
\item{\code{trainL}} {Number of time points used for conceptor training.}
\item{\code{error}} {Error of conceptor fit to data (<= tolerance).}
\item{\code{C}} {Conceptor matrices from each RNN fit.}
\item{\code{W}} {RNN matrices for each fit.}
\item{\code{ResStates}} {Time-ordered reservoir states for each RNN fit.}
}
}
}
\description{
Performs the conceptor change point algorithm to determine the location and significance of the most likely change point in a dependent, multivariate time series. Requires specification of a training length, or a combined washout and training length.
}
\details{
Provides an estimate of the most likely change point location in a multivariate time series. Fits a series of conceptor matrices to a representative training window of data, and compares the evolution of the RNN reservoir states to the original computed conceptor spaces. Method assumes that the training window is at least wide-sense cyclostationary, or there is not a long run trend present. The training window should be representative in the sense that it captures a full range of dynamics of the system. Change points are identified from a Kolmogorov-Smirnov like statistic based on a univariate sequence of derived cosine similarity measures. Significance estimates are obtained from a moving block bootstrap of the orginal data.
}
\examples{
\dontrun{
ccp(test_data, washL_plus_trainL = 150)
ccp(test_data, trainL = 100)
ccp(test_data, trainL = 100, tol = 0.08, nboots = 500)
}
}
