% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_integer}
\alias{check_integer}
\title{Check if value is a positive integer}
\usage{
check_integer(value)
}
\arguments{
\item{value}{Numeric value to check if positive integer.}
}
\value{
Boolean value (T/F) if value is a positive integer.
}
\description{
Check if value is a positive integer
}
