% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chooseMBBblockL}
\alias{chooseMBBblockL}
\title{Select optimal MBB block length}
\usage{
chooseMBBblockL(data, CRNNFit, kappa)
}
\arguments{
\item{data}{A T\code{x}d data set with columns as variables.}

\item{CRNNFit}{Output from fitCRNN function.}

\item{kappa}{Small constant that sets the change point domain (set as <= 0.01).}
}
\value{
Estimated optimal block length for MBB per Hall, Horowitz, and Jing (1995).
}
\description{
Selects optimal MBB block length with the Hall, Horowitz, and Jing (1995) algorithm.
}
