% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conceptorCalc}
\alias{conceptorCalc}
\title{Compute conceptor matrices}
\usage{
conceptorCalc(RS, aperture)
}
\arguments{
\item{RS}{reservoir states}

\item{aperture}{aperture value}
}
\value{
conceptor matrices
}
\description{
Compute conceptor matrices
}
