\name{conceptorCP-package}
\alias{conceptorCP-package}
\alias{conceptorCP}
\docType{package}
\title{
\packageTitle{conceptorCP}
}
\description{
\packageDescription{conceptorCP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{conceptorCP}
\packageIndices{conceptorCP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{conceptorCP}

Maintainer: \packageMaintainer{conceptorCP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
