% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CRNNBootstrap}
\alias{CRNNBootstrap}
\title{Bootstrap for CRNN}
\usage{
CRNNBootstrap(
  bootinput,
  W,
  C,
  RSInitial,
  washoutL,
  trainL,
  bscale,
  iscale,
  kappa
)
}
\arguments{
\item{bootinput}{bootstrapped data}

\item{W}{RNN matrices}

\item{C}{conceptor matrices}

\item{RSInitial}{initial reservoir state}

\item{washoutL}{length of washout data}

\item{trainL}{length of training data}

\item{bscale}{bias matrix scaling}

\item{iscale}{input matrix scaling}

\item{kappa}{small constant > 0}
}
\value{
bootstraped statistics
}
\description{
Bootstrap for CRNN
}
