% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_key_cust}
\alias{draw_key_cust}
\title{Adjust ggplot legend}
\usage{
draw_key_cust(data, params, size)
}
\arguments{
\item{data}{Data series to plot.}

\item{params}{Parameters of plot.}

\item{size}{Size of key.}
}
\value{
Adjusted plot settings.
}
\description{
Adjust ggplot legend
}
