% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitCRNN}
\alias{fitCRNN}
\title{Fit RNNs and conceptor matrices to data}
\usage{
fitCRNN(data, trainL, washoutL = "", tol = 0.04)
}
\arguments{
\item{data}{A T\code{x}d data set with columns as variables.}

\item{trainL}{Number of time points used for conceptor training.}

\item{washoutL}{Number of time points used for reservoir washout.}

\item{tol}{Error tolerance for conceptor fit to the data.}
}
\value{
List of RNN parameters and output from conceptor fit process.
}
\description{
Fits 100 RNNs to the data and computes the corresponding conceptor matrices. RNN parameters, initialized matrices, reservoir states, conceptor matrices, and cosine similarity measures between the reservoir states and conceptor spaces are stored.
}
