% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initRNN}
\alias{initRNN}
\title{Randomly initialize RNN matrices}
\usage{
initRNN(N, dimen)
}
\arguments{
\item{N}{RNN reservoir size.}

\item{dimen}{Dimension of original data, d.}
}
\value{
A N\code{x}(N+d+1) matrix of stacked RNN matrices in order: \cr
\[Wbias (N\code{x}1), Winput (N\code{x}d), Wres (N\code{x}N)\].
}
\description{
Randomly initializes matrices Wbias, Winput, and Wres for use in a recurrent neural network. Wbias and Winput are dense matrices with random normal entries and Wres is a sparse matrix with a spectral radius of 0.8.
}
