% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NRMSE}
\alias{NRMSE}
\title{Compute Normalized RMSE}
\usage{
NRMSE(output, target)
}
\arguments{
\item{output}{T\code{x}d series of interest}

\item{target}{T\code{x}d reference series}
}
\value{
average NRMSE of columns
}
\description{
Compute Normalized RMSE
}
