% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outputCalc}
\alias{outputCalc}
\title{Compute RNN output}
\usage{
outputCalc(Wout, CResStates)
}
\arguments{
\item{Wout}{RNN output matrix from \code{WoutCalc}}

\item{CResStates}{Reservoir states from \code{runCRNN}.}
}
\value{
Output from RNN.
}
\description{
Compute RNN output
}
