% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RNNParamFit}
\alias{RNNParamFit}
\title{Fit RNN parameters}
\usage{
RNNParamFit(input, W, RSInitial, washoutL, trainL, bscales, iscales)
}
\arguments{
\item{input}{data}

\item{W}{RNN matrices}

\item{RSInitial}{initial reservoir state}

\item{washoutL}{length of washout data}

\item{trainL}{length of trainind data}

\item{bscales}{bias matrix scaling vector}

\item{iscales}{input matrix scaling vector}
}
\value{
optimal RNN parameters
}
\description{
Fit RNN parameters
}
