% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runCRNN}
\alias{runCRNN}
\title{Run a recurrent neural network with conceptor matrix}
\usage{
runCRNN(input, W, C, RSInitial, washoutL, bscale, iscale)
}
\arguments{
\item{input}{A scaled T\code{x}d data set with variables as columns.}

\item{W}{RNN matrices output from \code{initRNN} function.}

\item{C}{Conceptor matrix(ces).}

\item{RSInitial}{Initial reservoir state.}

\item{washoutL}{Number of time points used for reservoir washout.}

\item{bscale}{Scale of RNN bias matrix.}

\item{iscale}{Scale of RNN input matrix.}
}
\value{
RNN reservoir states (controlled by a conceptor) for all T points.
}
\description{
Run a recurrent neural network with conceptor matrix
}
