% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runRNN}
\alias{runRNN}
\title{Run a RNN with specifications}
\usage{
runRNN(input, W, RSInitial, bscale, iscale)
}
\arguments{
\item{input}{data}

\item{W}{RNN matrices}

\item{RSInitial}{initial reservoir state}

\item{bscale}{bias matrix scaling}

\item{iscale}{input matrix scaling}
}
\value{
RNN reservoir states
}
\description{
Run a RNN with specifications
}
