% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scalein}
\alias{scalein}
\title{Scale columns of data to range (-1,1)}
\usage{
scalein(input_data)
}
\arguments{
\item{input_data}{A data set with columns as variables.}
}
\value{
The scaled data set with the same dimension as \code{input_data}.
}
\description{
Scales columns of data such that the middle 95% of the values are between -1 and 1 without changing the relative center.
}
