% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test data for use in examples
3-Dimensional VAR1 data set with chagne point at T = 500}
\format{
a tibble with 1000 rows and 3 variables (Series1, Series2, Series3)

set.seed(138)
CoeffMatrix1 <- matrix(c(0.6, -0.4, 0, -0.2, 0.8, -0.3, 0, 0.1, 0.3), nrow = 3)
CoeffMatrix2 <- matrix(c(0.2, -0.7, -0.2, 0, 0.9, -0.1, 0.2, 0.5, 0.1), nrow = 3)
RandInnovs <- matrix(rnorm(3000), ncol = 3)
test_data <- dplyr::tibble(Series1 = rnorm(1), Series2 = rnorm(1), Series3 = rnorm(1))
for(j in 1:500) {
   test_data[j + 1,] <- as.matrix(test_data[j,]) %*% CoeffMatrix1 + RandInnovs[j,]
}
for(j in 501:1000) {
   test_data[j + 1,] <- as.matrix(test_data[j,]) %*% CoeffMatrix2 + RandInnovs[j,]
}
test_data <- test_data[-1,]
}
\usage{
test_data
}
\description{
Test data for use in examples
3-Dimensional VAR1 data set with chagne point at T = 500
}
\keyword{datasets}
