% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WoutCalc}
\alias{WoutCalc}
\title{Compute output matrix}
\usage{
WoutCalc(input, ResStates, regular)
}
\arguments{
\item{input}{data}

\item{ResStates}{reservoir states}

\item{regular}{regularization parameter}
}
\value{
Wout matrix from ridge regression
}
\description{
Compute output matrix
}
