% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{angleCalc}
\alias{angleCalc}
\title{Compute cosine similarity between reservoir and conceptor}
\usage{
angleCalc(RS, CRS)
}
\arguments{
\item{RS}{Reservoir states from \code{runCRNN} before conceptor multiplication.}

\item{CRS}{Reservoir states from \code{runCRNN} after conceptor multiplication.}
}
\value{
Cosine similarities between reservoir and conceptor space at each time point.
}
\description{
Compute cosine similarity between reservoir and conceptor
}
