% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{angleCalc}
\alias{angleCalc}
\title{Compute cosine similarities}
\usage{
angleCalc(RS, CRS)
}
\arguments{
\item{RS}{RNN reservoir states}

\item{CRS}{conceptor multiplied reservoir states}
}
\value{
cosine similarity sequence
}
\description{
Compute cosine similarities
}
