% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chooseMBBblockL}
\alias{chooseMBBblockL}
\title{Title}
\usage{
chooseMBBblockL(data, CRNNFit)
}
\arguments{
\item{data}{A T\code{x}d data set with columns as variables.}

\item{CRNNFit}{Output from fitCRNN function.}
}
\value{
Estimated optimal block length for MBB per Hall, Horowitz, and Jung (1995).
}
\description{
Title
}
