% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conceptorCalc}
\alias{conceptorCalc}
\title{Compute conceptor matrices}
\usage{
conceptorCalc(RS, aperture)
}
\arguments{
\item{RS}{Reservoir states for conceptor computation.}

\item{aperture}{Aperture parameter.}
}
\value{
Conceptor matrix(ces).
}
\description{
Compute conceptor matrices
}
