% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CRNNBootstrap}
\alias{CRNNBootstrap}
\title{Run a RNN with conceptor matrix on bootstrapped data}
\usage{
CRNNBootstrap(bootinput, W, C, RSInitial, washoutL, trainL, bscale, iscale)
}
\arguments{
\item{bootinput}{A scaled T\code{x}d bootstrap data set with variables as columns.}

\item{W}{RNN matrices output from \code{initRNN} function.}

\item{C}{Conceptor matrix(ces).}

\item{RSInitial}{Initial reservoir state.}

\item{washoutL}{Number of time points used for reservoir washout.}

\item{trainL}{Number of time points used for conceptor training.}

\item{bscale}{Scale of RNN bias matrix.}

\item{iscale}{Scale of RNN input matrix.}
}
\value{
RNN reservoir states (controlled by a conceptor) for bootstrapped data.
}
\description{
Run a RNN with conceptor matrix on bootstrapped data
}
