% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KSstatCalc}
\alias{KSstatCalc}
\title{Compute Kolmogorov-Smirnov test statistic}
\usage{
KSstatCalc(angles)
}
\arguments{
\item{angles}{Cosine similarities from \code{angleCalc}.}
}
\value{
Kolmogorov-Smirnov like statistic at each time point.
}
\description{
Compute Kolmogorov-Smirnov test statistic
}
