% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KSstatCalc}
\alias{KSstatCalc}
\title{Compute CUSUM-like statistic}
\usage{
KSstatCalc(angles, kappa)
}
\arguments{
\item{angles}{cosine similarity sequence}

\item{kappa}{small constant > 0}
}
\value{
modified Kolmogorov-Smirnov test statistic
}
\description{
Compute CUSUM-like statistic
}
