% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NRMSE}
\alias{NRMSE}
\title{Calculate NRMSE}
\usage{
NRMSE(output, target)
}
\arguments{
\item{output}{observed data}

\item{target}{reference data}
}
\value{
NRMSE of observed data to target data
}
\description{
Calculate NRMSE
}
