% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NRMSE}
\alias{NRMSE}
\title{Normalized RMSE}
\usage{
NRMSE(output, target)
}
\arguments{
\item{output}{Data set of interest with columns as variables.}

\item{target}{Reference data set with same dimension as \code{output}.}
}
\value{
NRMSE of \code{output} relative to \code{target}.
}
\description{
Normalized RMSE
}
