% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outputCalc}
\alias{outputCalc}
\title{Compute RNN output}
\usage{
outputCalc(Wout, CResStates)
}
\arguments{
\item{Wout}{RNN output matrix from WoutCalc}

\item{CResStates}{conceptor multiplied reservoir states}
}
\value{
CRNN output
}
\description{
Compute RNN output
}
