% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotCP}
\alias{plotCP}
\title{Vizualize conceptor CP method}
\usage{
plotCP(ccp_output, nbreaks = 10)
}
\arguments{
\item{ccp_output}{Output from the conceptorCP function.}

\item{nbreaks}{Number of windows to divide series for visual.}
}
\value{
Plots in the following order:
\itemize{
\item \emph{Top}: Time-ordered series of statistics with most likely change point location and MBB results.
\item \emph{Middle}: Time-ordered cosine similarities between reservoir states and conceptor space.
\item \emph{Bottom}: Comparison of cosine similarity empirical CDFs (each window against the full time series).
}
}
\description{
Plots estimate and internal dynamics of the conceptor change point method.
}
\details{
Plots the time-ordered series of Kolmogorov-Smiornov like statistics from the conceptor change point method along with quantiles of the moving block bootstrap null distribution. Provides a visual aid of the relationship between the computed conceptor spaces and the propagating reservoir states over time with a cosine similarity measure. A comparison of the empirical disribution functions for windows of the cosine similarities are also included. Shading is relative and not on the same scale for all plots. Red shading represents time points where the dynamics are further away from the original conceptor space, and blue shading represents dynamics closer to the training window and the conceptor space.
}
\examples{
\donttest{
ccp_output <- ccp(test_data, washoutL_plus_trainL = 150)
plotCP(ccp_output)
}
}
