% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RNNParamFit}
\alias{RNNParamFit}
\title{Fit RNN to data}
\usage{
RNNParamFit(input, W, RSInitial, washoutL, trainL, bscales, iscales)
}
\arguments{
\item{input}{T\code{d} data}

\item{W}{cube of RNN matrices}

\item{RSInitial}{initial resercoir state}

\item{washoutL}{length of reservoir washout}

\item{trainL}{length of conceptor training}

\item{bscales}{bias matrix scalings}

\item{iscales}{input matrix scalings}
}
\value{
RNN parameters
}
\description{
Fit RNN to data
}
