% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RNNParamFit}
\alias{RNNParamFit}
\title{Fit RNN parameters to data}
\usage{
RNNParamFit(input, W, RSInitial, washoutL, trainL, bscales, iscales)
}
\arguments{
\item{input}{A scaled T\code{x}d data set with variables as columns.}

\item{W}{RNN matrices output from \code{initRNN} function.}

\item{RSInitial}{Initial reservoir state.}

\item{washoutL}{Number of time points used for reservoir washout.}

\item{trainL}{Number of time points used for conceptor training.}

\item{bscales}{Scalings for RNN bias matrix.}

\item{iscales}{Scalings for RNN input matrix.}
}
\value{
RNN parameters for best fit to data.
}
\description{
Fit RNN parameters to data
}
