% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runCRNN}
\alias{runCRNN}
\title{Run a RNN with conceptor matrix}
\usage{
runCRNN(input, W, C, RSInitial, washoutL, bscale, iscale)
}
\arguments{
\item{input}{data}

\item{W}{RNN matrices}

\item{C}{conceptor matrices}

\item{RSInitial}{initial reservoir state}

\item{washoutL}{length of washout}

\item{bscale}{bias matrix scaling}

\item{iscale}{input matrix scaling}
}
\value{
RNN reservoir states with conceptor control
}
\description{
Run a RNN with conceptor matrix
}
