% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runCRNN}
\alias{runCRNN}
\title{Propagate RNN with conceptors}
\usage{
runCRNN(input, W, C, RSInitial, washoutL, bscale, iscale)
}
\arguments{
\item{input}{T\code{x}d data}

\item{W}{cube of RNN matrices}

\item{C}{cube of conceptor matrices}

\item{RSInitial}{initial reservoir state}

\item{washoutL}{length of reservoir washout}

\item{bscale}{bias matrix scaling}

\item{iscale}{input matrix scaling}
}
\value{
conceptor multiplied reservoir states
}
\description{
Propagate RNN with conceptors
}
