% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runRNN}
\alias{runRNN}
\title{Run a recurrent neural network}
\usage{
runRNN(input, W, RSInitial, bscale, iscale)
}
\arguments{
\item{input}{A scaled T\code{x}d data set with variables as columns.}

\item{W}{RNN matrices output from \code{initRNN} function.}

\item{RSInitial}{Initial reservoir state.}

\item{bscale}{Scale of RNN bias matrix.}

\item{iscale}{Scale of RNN input matrix.}
}
\value{
RNN reservoir states for all T points.
}
\description{
Run a recurrent neural network
}
