% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runRNN}
\alias{runRNN}
\title{Propagate RNN}
\usage{
runRNN(input, W, RSInitial, bscale, iscale)
}
\arguments{
\item{input}{T\code{x}d data}

\item{W}{cube of RNN matrices}

\item{RSInitial}{initial reservoir state}

\item{bscale}{bias matrix scaling}

\item{iscale}{input matrix scaling}
}
\value{
RNN reservoir states
}
\description{
Propagate RNN
}
