% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test data for use in examples}
\format{
Tibble with 1000 rows and 3 variables (Series1, Series2, Series3)
}
\usage{
test_data
}
\description{
3-Dimensional VAR1 data set with change point at T = 500
}
\details{
Details of data creation:
\itemize{
\item Random seed set to 138 with \code{set.seed(138)}
\item VAR1 coefficient matrix for time points 1 - 500, \cr
\eqn{[0.6, 0, 0 // -0.2, 0.8, 0 // 0, 0.1, 0.4]}
\item VAR1 coefficient matrix for time points 501 - 1000, \cr
\eqn{[0, -0.8, -0 // 0.4, 0.2, 0 // 0.5, 0, -0.1]}
\item Random innovations generated with \code{matrix(rnorm(3000), ncol = 3)}
\item Each of the three variables initialized with \code{rnorm(1)}
}
}
\keyword{datasets}
