% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WoutCalc}
\alias{WoutCalc}
\title{Title Compute RNN output matrix}
\usage{
WoutCalc(input, ResStates, regular)
}
\arguments{
\item{input}{T\xode{x}d data}

\item{ResStates}{reservoir states from RNN}

\item{regular}{Ridge regression regularization parameter}
}
\value{
Wout output matrix
}
\description{
Title Compute RNN output matrix
}
