% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WoutCalc}
\alias{WoutCalc}
\title{Compute RNN output matrix}
\usage{
WoutCalc(input, ResStates, regular)
}
\arguments{
\item{input}{A scaled T\code{x}d data set with variables as columns.}

\item{ResStates}{Reservoir states from \code{runCRNN}.}

\item{regular}{Regularization parameter for ridge regression.}
}
\value{
RNN output matrix(ces).
}
\description{
Compute RNN output matrix
}
